+++ src/docs/asciidoc/api/balance/balance.adoc
@@ -9,14 +9,14 @@ include::{snippets}/get-balance/path-parameters.adoc[]
 include::{snippets}/get-balance/http-response.adoc[]
 include::{snippets}/get-balance/response-fields.adoc[]
 
-[[update-balance]]
+[[charge-balance]]
 === 잔액 충전
 
 ==== HTTP Request
-include::{snippets}/update-balance/http-request.adoc[]
-include::{snippets}/update-balance/path-parameters.adoc[]
-include::{snippets}/update-balance/request-fields.adoc[]
+include::{snippets}/charge-balance/http-request.adoc[]
+include::{snippets}/charge-balance/path-parameters.adoc[]
+include::{snippets}/charge-balance/request-fields.adoc[]
 
 ==== HTTP Response
-include::{snippets}/update-balance/http-response.adoc[]
-include::{snippets}/update-balance/response-fields.adoc[]
+include::{snippets}/charge-balance/http-response.adoc[]
+include::{snippets}/charge-balance/response-fields.adoc[]
