+++ src/docs/asciidoc/api/product/product.adoc
@@ -7,3 +7,13 @@ include::{snippets}/get-products/http-request.adoc[]
 ==== HTTP Response
 include::{snippets}/get-products/http-response.adoc[]
 include::{snippets}/get-products/response-fields.adoc[]
+
+[[get-ranks]]
+=== 상위 상품 Top5 목록 조회
+
+==== HTTP Request
+include::{snippets}/get-ranks/http-request.adoc[]
+
+==== HTTP Response
+include::{snippets}/get-ranks/http-response.adoc[]
+include::{snippets}/get-ranks/response-fields.adoc[]
