--- src/docs/asciidoc/api/rank/rank.adoc
@@ -1,9 +1,0 @@
-[[get-ranks]]
-=== 상위 상품 Top5 목록 조회
-
-==== HTTP Request
-include::{snippets}/get-ranks/http-request.adoc[]
-
-==== HTTP Response
-include::{snippets}/get-ranks/http-response.adoc[]
-include::{snippets}/get-ranks/response-fields.adoc[]
