--- src/main/java/kr/hhplus/be/ecommerce/api/controller/ApiResponse.java
@@ -1,39 +1,0 @@
-package kr.hhplus.be.ecommerce.api.controller;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
-
-@Getter
-public class ApiResponse<T> {
-
-    private final int code;
-    private final String message;
-
-    @JsonInclude(NON_NULL)
-    private final T data;
-
-    private ApiResponse(int code, String message, T data) {
-        this.code = code;
-        this.message = message;
-        this.data = data;
-    }
-
-    public static <T> ApiResponse<T> success(T data) {
-        return new ApiResponse<>(
-            HttpStatus.OK.value(),
-            HttpStatus.OK.getReasonPhrase(),
-            data
-        );
-    }
-
-    public static <T> ApiResponse<T> success() {
-        return success(null);
-    }
-
-    public static <T> ApiResponse<T> fail(int code, String message) {
-        return new ApiResponse<>(code, message, null);
-    }
-}
