+++ src/main/java/kr/hhplus/be/ecommerce/interfaces/order/OrderController.java
@@ -1,0 +1,22 @@
+package kr.hhplus.be.ecommerce.interfaces.order;
+
+import jakarta.validation.Valid;
+import kr.hhplus.be.ecommerce.application.order.OrderFacade;
+import kr.hhplus.be.ecommerce.interfaces.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class OrderController {
+
+    private final OrderFacade orderFacade;
+
+    @PostMapping("/api/v1/orders")
+    public ApiResponse<Void> createOrder(@Valid @RequestBody OrderRequest.Create request) {
+        orderFacade.orderPayment(request.toCriteria());
+        return ApiResponse.success();
+    }
+}
