+++ src/test/java/kr/hhplus/be/ecommerce/docs/order/OrderControllerDocsTest.java
@@ -1,9 +1,9 @@
 package kr.hhplus.be.ecommerce.docs.order;
 
-import kr.hhplus.be.ecommerce.api.controller.order.OrderController;
-import kr.hhplus.be.ecommerce.api.controller.order.request.OrderCreateRequest;
-import kr.hhplus.be.ecommerce.api.controller.order.request.OrderProductRequest;
+import kr.hhplus.be.ecommerce.application.order.OrderFacade;
 import kr.hhplus.be.ecommerce.docs.RestDocsSupport;
+import kr.hhplus.be.ecommerce.interfaces.order.OrderController;
+import kr.hhplus.be.ecommerce.interfaces.order.OrderRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
@@ -11,6 +11,7 @@ import org.springframework.restdocs.payload.JsonFieldType;
 
 import java.util.List;
 
+import static org.mockito.Mockito.mock;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
@@ -20,20 +21,22 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 class OrderControllerDocsTest extends RestDocsSupport {
 
+    private final OrderFacade orderFacade = mock(OrderFacade.class);
+
     @Override
     protected Object initController() {
-        return new OrderController();
+        return new OrderController(orderFacade);
     }
 
     @DisplayName("주문/결제 완료 API")
     @Test
     void createOrder() throws Exception {
         // given
-        OrderCreateRequest request = OrderCreateRequest.of(
+        OrderRequest.Create request = OrderRequest.Create.of(
             1L,
             1L,
             List.of(
-                OrderProductRequest.of(1L, 2)
+                OrderRequest.Product.of(1L, 2)
             )
         );
